requirements.txt
fastapi==0.128.0
matplotlib==3.7.2
numpy==1.23.0
opencv_python==4.8.0.74
opencv_python_headless==4.11.0.86
Pillow==9.5.0
Pillow==12.0.0
ultralytics==8.3.28
uvicorn==0.40.0
